Main file and regression tests 1.0
To analyze a written code, call 'setlx analyzeCode.stlx --params x'
where x is the file(s) you would like to analyze using the command
prompt in windows or the terminator in mac or linux.
Also in case any amendments were added to the original file make sure to
change both the main file and the other one in the Testing/ directory.
To check that the amended code didn't do any harm to the code run the
regression tests using the command 'setlx regressionTests.stlx' in the
Testing/ directory.